+++ flixopt/components.py
@@ -811,7 +811,7 @@ class TransmissionModel(ComponentModel):
 
     def __init__(self, model: FlowSystemModel, element: Transmission):
         if (element.absolute_losses is not None) and np.any(element.absolute_losses != 0):
-            for flow in element.inputs + element.outputs:
+            for flow in element.flows.values():
                 if flow.status_parameters is None:
                     flow.status_parameters = StatusParameters()
                     flow.status_parameters.link_to_flow_system(
@@ -877,8 +877,8 @@ class LinearConverterModel(ComponentModel):
 
         # Create conversion factor constraints if specified
         if self.element.conversion_factors:
-            all_input_flows = set(self.element.inputs)
-            all_output_flows = set(self.element.outputs)
+            all_input_flows = set(self.element.inputs.values())
+            all_output_flows = set(self.element.outputs.values())
 
             # für alle linearen Gleichungen:
             for i, conv_factors in enumerate(self.element.conversion_factors):